requirements.txt
fastapi[standard]
twilio
python-dotenv
SQLAlchemy
openai
pydub
pandas
google-cloud-storage
uvicorn
openpyxl